ecom_project1/nk_mart/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-payments==2.0.0
django-payments-razorpay==1.0
django-phonenumber-field==7.3.0
django-shopping-cart==0.1
gunicorn==21.2.0
idna==3.6
packaging==24.0
phonenumberslite==8.13.32
pillow==10.2.0
razorpay==1.4.1
requests==2.31.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
